Reafactor: unit test split the same way as keyword_importer and keyword_extractor
