+++ tests/__init__.py
@@ -4,6 +4,7 @@ from tests.unit.api.endpoints.keywords_tests import KeywordsApiTest
 from tests.unit.api.endpoints.statistics_tests import StatisticsApiTest
 from tests.unit.api.endpoints.version_tests import VersionApiTest
 from tests.unit.cli.api_client import ApiClientTests
+from tests.unit.cli.keywords.keywords_extractor import KeywordsExtractorTests
 from tests.unit.cli.keywords.keywords_importer import KeywordsImporterTests
 from tests.unit.cli.statistics.statistics_extractor import StatisticsExtractorTests
 from tests.unit.db.collection_repo_tests import CollectionRepositoryTest
