+++ tests/unit/cli/keywords/keywords_importer.py
@@ -1,212 +1,10 @@
 import copy
 import responses
 import unittest
-from pathlib import Path
-from robot.libdocpkg import LibraryDocumentation
-import robot.libraries
 
-from rfhub2.cli.keywords.keywords_importer import (
-    CollectionUpdateWithKeywords,
-    KeywordsImporter,
-)
+from rfhub2.cli.keywords.keywords_importer import KeywordsImporter
 from rfhub2.cli.api_client import Client
-from rfhub2.model import Collection, CollectionUpdate, KeywordUpdate, NestedKeyword
-
-FIXTURE_PATH = Path.cwd() / "tests" / "fixtures" / "initial"
-STATISTICS_PATH = FIXTURE_PATH / ".." / "statistics"
-EXPECTED_LIBDOC = {
-    "doc": "Documentation for library ``Test Libdoc File``.",
-    "doc_format": "ROBOT",
-    "name": "Test Libdoc File",
-    "scope": "GLOBAL",
-    "type": "library",
-    "version": "3.2.0",
-    "keywords": [{"name": "Someone Shall Pass", "args": '["who"]', "doc": ""}],
-}
-EXPECTED_INIT_DOC = """\n\nHere goes some docs that should appear on rfhub2 if init is parametrised
-\nThe library import:
-\nExamples:
-| Library    LibWithInit   dummy=../one               # add one dummy
-| Library    LibWithInit   path=../one,/global        # add two dummies"""
-EXPECTED_KEYWORDS = [
-    KeywordUpdate(
-        args="",
-        doc="This keyword was imported from file\n"
-        "with .resource extension, available since RFWK 3.1",
-        name="Keyword 1 Imported From Resource File",
-        tags=["first_tag"],
-    ),
-    KeywordUpdate(
-        args='["arg_1", "arg_2"]',
-        doc="This keyword was imported from file\n"
-        "with .resource extension, available since RFWK 3.1",
-        name="Keyword 2 Imported From Resource File",
-        tags=["first_tag", "second_tag"],
-    ),
-]
-EXPECTED_TRAVERSE_PATHS_INIT = {FIXTURE_PATH / "LibWithInit"}
-EXPECTED_TRAVERSE_PATHS_NO_INIT = {
-    FIXTURE_PATH / "LibsWithEmptyInit" / "LibWithEmptyInit1.py",
-    FIXTURE_PATH / "LibsWithEmptyInit" / "LibWithEmptyInit2.py",
-}
-EXPECTED_GET_LIBRARIES = (
-    EXPECTED_TRAVERSE_PATHS_INIT
-    | EXPECTED_TRAVERSE_PATHS_NO_INIT
-    | {
-        FIXTURE_PATH / "SingleClassLib" / "SingleClassLib.py",
-        FIXTURE_PATH / "test_libdoc_file.xml",
-        FIXTURE_PATH / "test_resource.resource",
-        FIXTURE_PATH / "test_robot.robot",
-        FIXTURE_PATH / "arg_parse.py",
-        FIXTURE_PATH / "data_error.py",
-        FIXTURE_PATH / "LibWithInit" / "test_res_lib_dir.resource",
-    }
-)
-EXPECTED_GET_EXECUTION_PATHS = {
-    STATISTICS_PATH / "output.xml",
-    STATISTICS_PATH / "subdir" / "output.xml",
-}
-EXPECTED_COLLECTION = CollectionUpdate(
-    doc="Overview that should be imported for SingleClassLib.",
-    doc_format="ROBOT",
-    name="SingleClassLib",
-    path=str(FIXTURE_PATH / "SingleClassLib" / "SingleClassLib.py"),
-    scope="TEST",
-    type="LIBRARY",
-    version="1.2.3",
-)
-
-EXPECTED_COLLECTION_KEYWORDS_1_1 = KeywordUpdate(
-    args="",
-    doc="Docstring for single_class_lib_method_1",
-    name="Single Class Lib Method 1",
-    tags=["tag_1", "tag_2"],
-)
-EXPECTED_COLLECTION_KEYWORDS_1_2 = KeywordUpdate(
-    args="",
-    doc="Docstring for single_class_lib_method_2",
-    name="Single Class Lib Method 2",
-    tags=[],
-)
-EXPECTED_COLLECTION_KEYWORDS_1_3 = KeywordUpdate(
-    args='["param_1", "param_2"]',
-    doc="Docstring for single_class_lib_method_3 with two params",
-    name="Single Class Lib Method 3",
-    tags=[],
-)
-EXPECTED_COLLECTION_KEYWORDS_1 = [
-    EXPECTED_COLLECTION_KEYWORDS_1_1,
-    EXPECTED_COLLECTION_KEYWORDS_1_2,
-    EXPECTED_COLLECTION_KEYWORDS_1_3,
-]
-EXISTING_COLLECTION_KEYWORDS = [
-    NestedKeyword(**{**EXPECTED_COLLECTION_KEYWORDS_1_3.dict(), "id": 1})
-]
-EXISTING_COLLECTION = Collection(
-    **{
-        **EXPECTED_COLLECTION.dict(),
-        "id": 1,
-        "keywords": [
-            NestedKeyword(**{**EXPECTED_COLLECTION_KEYWORDS_1_3.dict(), "id": 1})
-        ],
-    }
-)
-EXPECTED_COLLECTION_2 = CollectionUpdate(
-    doc="Documentation for library ``Test Libdoc File``.",
-    doc_format="ROBOT",
-    name="Test Libdoc File",
-    path=str(FIXTURE_PATH / "test_libdoc_file.xml"),
-    scope="GLOBAL",
-    type="LIBRARY",
-    version="3.2.0",
-)
-EXPECTED_COLLECTION_KEYWORDS_2_1 = KeywordUpdate(
-    args='["who"]', doc="", name="Someone Shall Pass", tags=[]
-)
-EXPECTED_COLLECTION_KEYWORDS_2 = [EXPECTED_COLLECTION_KEYWORDS_2_1]
-EXISTING_COLLECTION_2 = Collection(
-    **{
-        **EXPECTED_COLLECTION_2.dict(),
-        "id": 1,
-        "keywords": [
-            NestedKeyword(**{**EXPECTED_COLLECTION_KEYWORDS_2_1.dict(), "id": 1})
-        ],
-    }
-)
-EXPECTED_ADD_COLLECTIONS = [{"name": "Test Libdoc File", "keywords": 1}]
-EXPECTED_UPDATE_COLLECTIONS = [
-    {"name": "a", "keywords": 1},
-    {"name": "b", "keywords": 1},
-    {"name": "c", "keywords": 1},
-    {"name": "d", "keywords": 1},
-    {"name": "e", "keywords": 1},
-]
-KEYWORDS_1 = [
-    {
-        "args": "",
-        "doc": "Docstring for single_class_lib_method_1",
-        "name": "Single Class Lib Method 1",
-    },
-    {
-        "args": "",
-        "doc": "Docstring for single_class_lib_method_2",
-        "name": "Single Class Lib Method 2",
-    },
-    {
-        "args": '["param_1", "param_2"]',
-        "doc": "Docstring for single_class_lib_method_3 with two params",
-        "name": "Single Class Lib Method 3",
-    },
-]
-KEYWORDS_2 = [{"args": '["who"]', "doc": "", "name": "Someone Shall Pass"}]
-KEYWORDS_EXTENDED = [
-    {
-        "args": "",
-        "doc": "Docstring for single_class_lib_method_1",
-        "name": "Single Class Lib Method 1",
-        "id": 15,
-        "synopsis": "Docstring for lib_with_empty_init_1_method_1",
-        "html_doc": "<p>Docstring for lib_with_empty_init_1_method_1</p>",
-        "arg_string": "",
-    },
-    {
-        "args": "",
-        "doc": "Docstring for single_class_lib_method_2",
-        "name": "Single Class Lib Method 2",
-        "id": 16,
-        "synopsis": "Docstring for lib_with_empty_init_1_method_1",
-        "html_doc": "<p>Docstring for lib_with_empty_init_1_method_1</p>",
-        "arg_string": "",
-    },
-    {
-        "args": '["param_1", "param_2"]',
-        "doc": "Docstring for single_class_lib_method_3 with two params",
-        "name": "Single Class Lib Method 3",
-        "id": 17,
-        "synopsis": "Docstring for lib_with_empty_init_1_method_1",
-        "html_doc": "<p>Docstring for lib_with_empty_init_1_method_1</p>",
-        "arg_string": "",
-    },
-]
-
-EXPECTED_BUILT_IN_LIBS = {
-    Path(robot.libraries.__file__).parent / "BuiltIn.py",
-    Path(robot.libraries.__file__).parent / "Collections.py",
-    Path(robot.libraries.__file__).parent / "DateTime.py",
-    Path(robot.libraries.__file__).parent / "Easter.py",
-    Path(robot.libraries.__file__).parent / "OperatingSystem.py",
-    Path(robot.libraries.__file__).parent / "Process.py",
-    Path(robot.libraries.__file__).parent / "Screenshot.py",
-    Path(robot.libraries.__file__).parent / "String.py",
-    Path(robot.libraries.__file__).parent / "Telnet.py",
-    Path(robot.libraries.__file__).parent / "XML.py",
-}
-EXPECTED_COLLECTION_WITH_KW_1 = CollectionUpdateWithKeywords(
-    EXPECTED_COLLECTION, EXPECTED_COLLECTION_KEYWORDS_1
-)
-EXPECTED_COLLECTION_WITH_KW_2 = CollectionUpdateWithKeywords(
-    EXPECTED_COLLECTION_2, EXPECTED_COLLECTION_KEYWORDS_2
-)
+from .test_data import *
 
 
 class KeywordsImporterTests(unittest.TestCase):
@@ -461,67 +259,6 @@ class KeywordsImporterTests(unittest.TestCase):
             existing_collection.keywords = []
             self.assertListEqual([existing_collection], result)
 
-    def test_traverse_paths_should_return_set_of_path_on_lib_with_init(self):
-        result = self.rfhub_importer._traverse_paths(self.fixture_path / "LibWithInit")
-        self.assertEqual(result, EXPECTED_TRAVERSE_PATHS_INIT)
-
-    def test_traverse_paths_should_return_set_of_paths_on_libs_with_empty_init(self):
-        result = self.rfhub_importer._traverse_paths(
-            self.fixture_path / "LibsWithEmptyInit"
-        )
-        self.assertEqual(result, EXPECTED_TRAVERSE_PATHS_NO_INIT)
-
-    def test_get_libraries_paths_should_return_set_of_paths(self):
-        result = self.rfhub_importer.get_libraries_paths()
-        self.assertEqual(result, EXPECTED_GET_LIBRARIES)
-
-    def test_get_libraries_paths_should_return_set_of_paths_on_installed_keywords(self):
-        self.rfhub_importer = KeywordsImporter(self.client, tuple(), False, False)
-        result = self.rfhub_importer.get_libraries_paths()
-        self.assertEqual(result, EXPECTED_BUILT_IN_LIBS)
-
-    def test_get_libraries_paths_should_return_set_of_paths_when_paths_are_tuple(self):
-        self.rfhub_importer = KeywordsImporter(
-            self.client,
-            (
-                self.fixture_path / "LibWithInit",
-                self.fixture_path / "LibsWithEmptyInit",
-            ),
-            True,
-            False,
-        )
-        result = self.rfhub_importer.get_libraries_paths()
-        self.assertEqual(
-            result, EXPECTED_TRAVERSE_PATHS_INIT | EXPECTED_TRAVERSE_PATHS_NO_INIT
-        )
-
-    def test__create_collections_should_return_collection_list(self):
-        result = self.rfhub_importer.create_collections(
-            {
-                FIXTURE_PATH / "SingleClassLib" / "SingleClassLib.py",
-                FIXTURE_PATH / "test_libdoc_file.xml",
-            }
-        )
-        self.assertCountEqual(
-            result, [EXPECTED_COLLECTION_WITH_KW_1, EXPECTED_COLLECTION_WITH_KW_2]
-        )
-
-    def test_create_collection_should_return_collection(self):
-        result = self.rfhub_importer.create_collection(
-            FIXTURE_PATH / "SingleClassLib" / "SingleClassLib.py"
-        )
-        self.assertEqual(EXPECTED_COLLECTION_WITH_KW_1, result)
-
-    def test_create_collections_should_return_empty_list_on_data_error(self):
-        result = self.rfhub_importer.create_collections(
-            {FIXTURE_PATH / "data_error.py"}
-        )
-        self.assertListEqual([], result)
-
-    def test_create_collections_should_return_empty_list_on_system_exit(self):
-        result = self.rfhub_importer.create_collections({FIXTURE_PATH / "arg_parse.py"})
-        self.assertListEqual([], result)
-
     def test_update_collections_should_insert_collections(self):
         existing_collections = [
             Collection(
@@ -685,171 +422,6 @@ class KeywordsImporterTests(unittest.TestCase):
                     },
                 )
                 self.rfhub_importer.add_collections([EXPECTED_COLLECTION_WITH_KW_2])
-
-    def test_is_library_with_init_should_return_true_on_library_with_init(self):
-        file = self.fixture_path / "LibWithInit"
-        result = self.rfhub_importer._is_library_with_init(file)
-        self.assertTrue(
-            result, "method should return true if file is python library with init"
-        )
-
-    def test_is_library_with_init_should_return_false_on_library_without_init(self):
-        file = self.fixture_path / "LibsWithEmptyInit"
-        result = self.rfhub_importer._is_library_with_init(file)
-        self.assertFalse(
-            result, "method should return false if file is python library without init"
-        )
-
-    def test_is_robot_keyword_file_should_return_true_on_library(self):
-        file = self.fixture_path / "SingleClassLib" / "SingleClassLib.py"
-        result = self.rfhub_importer._is_robot_keyword_file(file)
-        self.assertTrue(result, "method should return true if file is python library")
-
-    def test_is_robot_keyword_file_should_return_true_on_libdoc(self):
-        file = self.fixture_path / "test_libdoc_file.xml"
-        result = self.rfhub_importer._is_robot_keyword_file(file)
-        self.assertTrue(result, "method should return true if file is libdoc file")
-
-    def test_is_robot_keyword_file_should_return_true_on_resource(self):
-        file = self.fixture_path / "test_resource.resource"
-        result = self.rfhub_importer._is_robot_keyword_file(file)
-        self.assertTrue(result, "method should return true if file is robot resource")
-
-    def test_is_library_file_should_return_false_on_lib_with_init(self):
-        file = self.fixture_path / "LibWithInit" / "__init__.py"
-        result = KeywordsImporter._is_library_file(file)
-        self.assertFalse(result, "method should return true if file is python library")
-
-    def test_is_library_file_should_return_false_on_library_with_init(self):
-        file = self.fixture_path / "LibWithInit" / "__init__.py"
-        result = KeywordsImporter._is_library_file(file)
-        self.assertFalse(
-            result, "method should return false if file is python library with init"
-        )
-
-    def test_is_libdoc_file_should_return_true_on_libdoc(self):
-        file = self.fixture_path / "test_libdoc_file.xml"
-        result = KeywordsImporter._is_libdoc_file(file)
-        self.assertTrue(result, "method should return true if file is libdoc file")
-
-    def test_is_libdoc_file_should_return_false_on_non_libdoc(self):
-        file = self.fixture_path / "not_libdoc_file.xml"
-        result = KeywordsImporter._is_libdoc_file(file)
-        self.assertFalse(
-            result, "method should return false if file is not libdoc file"
-        )
-
-    def test_is_libdoc_file_should_return_false_on_non_xml(self):
-        file = self.fixture_path / "_private_library.py"
-        result = KeywordsImporter._is_libdoc_file(file)
-        self.assertFalse(
-            result, "method should return false if file is not libdoc file"
-        )
-
-    def test_should_ignore_should_return_true_on_deprecated(self):
-        file = self.fixture_path / "deprecated_library.py"
-        result = KeywordsImporter._should_ignore(file)
-        self.assertTrue(
-            result, 'method should return true if file starts with "deprecated"'
-        )
-
-    def test_should_ignore_should_return_true_on_private(self):
-        file = self.fixture_path / "_private_library.py"
-        result = KeywordsImporter._should_ignore(file)
-        self.assertTrue(result, 'method should return true if file starts with "_"')
-
-    def test_should_ignore_should_return_true_on_excluded(self):
-        file = self.fixture_path / "remote.py"
-        result = KeywordsImporter._should_ignore(file)
-        self.assertTrue(
-            result, "method should return true if file in EXCLUDED_LIBRARIES"
-        )
-
-    def test_should_ignore_should_return_false_on_library_to_import(self):
-        file = self.fixture_path / "SingleClassLib" / "SingleClassLib.py"
-        result = KeywordsImporter._should_ignore(file)
-        self.assertFalse(
-            result, "method should return false if file should be imported"
-        )
-
-    def test_is_resource_file_should_return_true(self):
-        file = self.fixture_path / "test_resource.resource"
-        result = KeywordsImporter._is_resource_file(file)
-        self.assertTrue(result, "method should return true if file is resource file")
-
-    def test_is_resource_file_should_return_false(self):
-        file = self.fixture_path / "test_file_with_tests.robot"
-        result = KeywordsImporter._is_resource_file(file)
-        self.assertFalse(
-            result, "method should return false if file is not resource file"
-        )
-
-    def test_is_resource_file_should_return_false_on_init(self):
-        file = self.fixture_path / "__init__.robot"
-        result = KeywordsImporter._is_resource_file(file)
-        self.assertFalse(
-            result, "method should return false if file is not resource file"
-        )
-
-    def test_has_keyword_table_should_return_true(self):
-        data = "*** Keywords ***"
-        result = KeywordsImporter._has_keyword_table(data=data)
-        self.assertTrue(result, "method should return true if Keywords were found")
-
-    def test_has_keyword_table_should_return_false(self):
-        data = "*** Keys ***"
-        result = KeywordsImporter._has_keyword_table(data=data)
-        self.assertFalse(
-            result, "method should return false if Keywords were not found"
-        )
-
-    def test_has_test_case_table_should_return_true(self):
-        data = "*** Test Case ***"
-        result = KeywordsImporter._has_test_case_table(data=data)
-        self.assertTrue(result, "method should return true if Test Case were found")
-
-    def test_has_test_case_table_should_return_false(self):
-        data = "*** Test ***"
-        result = KeywordsImporter._has_test_case_table(data=data)
-        self.assertFalse(
-            result, "method should return false if Test Case were not found"
-        )
-
-    def test_serialise_libdoc_should_return_collection(self):
-        file = self.fixture_path / "test_libdoc_file.xml"
-        libdoc_1 = LibraryDocumentation(file)
-        serialised_libdoc = self.rfhub_importer._serialise_libdoc(libdoc_1, str(file))
-        self.assertEqual(serialised_libdoc, EXPECTED_COLLECTION_2)
-
-    def test_serialise_keywords_should_return_keywords(self):
-        file = self.fixture_path / "test_resource.resource"
-        libdoc_2 = LibraryDocumentation(file)
-        serialised_keywords = self.rfhub_importer._serialise_keywords(libdoc_2)
-        self.assertEqual(serialised_keywords, EXPECTED_KEYWORDS)
-
-    def test_extract_doc_from_libdoc_inits_should_return_doc_from_init(self):
-        file = str(self.fixture_path / "LibWithInit")
-        libdoc = LibraryDocumentation(file)
-        init_doc = self.rfhub_importer._extract_doc_from_libdoc_inits(libdoc.inits)
-        self.assertEqual(init_doc, EXPECTED_INIT_DOC)
-
-    def test_extract_doc_from_libdoc_inits_should_return_empty_string(self):
-        file = str(self.fixture_path / "SingleClassLib" / "SingleClassLib.py")
-        libdoc = LibraryDocumentation(file)
-        init_doc = self.rfhub_importer._extract_doc_from_libdoc_inits(libdoc.inits)
-        self.assertEqual(init_doc, "")
-
-    def test_robot_files_candidates_should_return_true_if_robot_files_present(self):
-        self.assertTrue(
-            self.rfhub_importer._robot_files_candidates(
-                self.fixture_path / "LibWithInit"
-            )
-        )
-        self.assertFalse(
-            self.rfhub_importer._robot_files_candidates(
-                self.fixture_path / "LibWithEmptyInit"
-            )
-        )
 
     def test_collection_path_and_name_match_should_return_true_when_matched(self):
         result = KeywordsImporter._collection_path_and_name_match(
